updating

+++ scripts/set-vault-password.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 # use this to pass the vault secret to ansible navigator
-yoursecret=12345678
+yoursecret=jgk3pr64
 while getopts "d" OPTION
 do
   case "$OPTION" in
@@ -13,11 +13,14 @@ done
 main()
 {
   # code goes here
-  echo -e '#!/bin/sh\necho ${ANSIBLE_VAULT_SECRET}' >> ~/.vault_password.sh
-  chmod 700 ~/.vault_password.sh
-  ln ~/.vault_password.sh .
+  if [[ ! -e ~/.vault_password.sh ]]
+  then
+    echo -e '#!/bin/sh\necho ${ANSIBLE_VAULT_SECRET}' >> ~/.vault_password.sh
+    chmod 700 ~/.vault_password.sh
+    ln ~/.vault_password.sh .
+  fi
   export ANSIBLE_VAULT_SECRET=${yoursecret}
-  ANSIBLE_VAULT_PASSWORD_FILE=.vault_password.sh
+  export ANSIBLE_VAULT_PASSWORD_FILE=.vault_password.sh
 
 }
